Comparing with https://github.com/kcp-dev/kubernetes/pull/34:

Before:
```
$ go test -bench=Mutation -run=LKSHDJSH ./test/e2e -count 10
BenchmarkMutation-8   	     786	   1954867 ns/op
BenchmarkMutation-8   	     751	   1795782 ns/op
BenchmarkMutation-8   	     768	   1846943 ns/op
BenchmarkMutation-8   	     706	   1538904 ns/op
BenchmarkMutation-8   	     596	   2106487 ns/op
BenchmarkMutation-8   	     652	   1935539 ns/op
BenchmarkMutation-8   	     502	   2275392 ns/op
BenchmarkMutation-8   	     548	   2194643 ns/op
BenchmarkMutation-8   	     765	   1598261 ns/op
BenchmarkMutation-8   	     648	   2263218 ns/op
PASS
ok  	github.com/kcp-dev/kcp/test/e2e	152.365s
```

After:

```
$ go test -bench=Mutation -run=LKSHDJSH ./test/e2e -count 10
goos: linux
goarch: amd64
pkg: github.com/kcp-dev/kcp/test/e2e
cpu: Intel(R) Core(TM) i7-8665U CPU @ 1.90GHz
BenchmarkMutation-8   	     894	   1467171 ns/op
BenchmarkMutation-8   	     693	   1672650 ns/op
BenchmarkMutation-8   	     793	   1596037 ns/op
BenchmarkMutation-8   	     843	   1571645 ns/op
BenchmarkMutation-8   	     604	   1858038 ns/op
BenchmarkMutation-8   	     601	   1984385 ns/op
BenchmarkMutation-8   	     853	   2079691 ns/op
BenchmarkMutation-8   	     421	   2667725 ns/op
BenchmarkMutation-8   	     572	   2504347 ns/op
BenchmarkMutation-8   	     660	   2072154 ns/op
PASS
ok  	github.com/kcp-dev/kcp/test/e2e	151.453s
```

Diff:

```
$ benchstat /tmp/old.txt /tmp/new.txt 
name        old time/op  new time/op  delta
Mutation-8  1.95ms ±21%  1.95ms ±37%   ~     (p=0.796 n=10+10)
```